Moved view parameters and widgets into their own modules
